Issue #4919 - make the SessionTracker the one to implement Graceful shutdown
