+++ jetty-websocket/websocket-javax-common/src/main/java/org/eclipse/jetty/websocket/javax/common/JavaxWebSocketContainer.java
@@ -23,7 +23,6 @@ import java.util.ArrayList;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
-import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.Executor;
 import java.util.function.Consumer;
 import javax.websocket.Extension;
@@ -32,15 +31,13 @@ import javax.websocket.WebSocketContainer;
 import org.eclipse.jetty.io.ByteBufferPool;
 import org.eclipse.jetty.util.DecoratedObjectFactory;
 import org.eclipse.jetty.util.component.ContainerLifeCycle;
-import org.eclipse.jetty.util.component.Graceful;
 import org.eclipse.jetty.websocket.core.Configuration;
 import org.eclipse.jetty.websocket.core.WebSocketComponents;
 import org.eclipse.jetty.websocket.core.WebSocketExtensionRegistry;
-import org.eclipse.jetty.websocket.util.ShutdownUtil;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-public abstract class JavaxWebSocketContainer extends ContainerLifeCycle implements javax.websocket.WebSocketContainer, Graceful
+public abstract class JavaxWebSocketContainer extends ContainerLifeCycle implements javax.websocket.WebSocketContainer
 {
     private static final Logger LOG = LoggerFactory.getLogger(JavaxWebSocketContainer.class);
     private final List<JavaxWebSocketSessionListener> sessionListeners = new ArrayList<>();
@@ -52,6 +49,7 @@ public abstract class JavaxWebSocketContainer extends ContainerLifeCycle impleme
     {
         this.components = components;
         addSessionListener(sessionTracker);
+        addBean(sessionTracker);
     }
 
     public abstract Executor getExecutor();
@@ -199,24 +197,5 @@ public abstract class JavaxWebSocketContainer extends ContainerLifeCycle impleme
                 LOG.info("Exception while invoking listener " + listener, x);
             }
         }
-    }
-
-    @Override
-    protected void doStart() throws Exception
-    {
-        sessionTracker.start();
-        super.doStart();
-    }
-
-    @Override
-    public CompletableFuture<Void> shutdown()
-    {
-        return ShutdownUtil.stop(sessionTracker);
-    }
-
-    @Override
-    public boolean isShutdown()
-    {
-        return sessionTracker.isStopped();
     }
 }
